+++ src/components/text-input.tsx
@@ -12,7 +12,7 @@ export default function TextInput(
       style={styles}
       onChange={props.onChange}
       className={
-        "h-12 min-w-12 bg-stone-800 rounded-lg text-white font-medium px-3 focus:outline-stone-400 focus:outline-2 inset-shadow-md " +
+        "h-12 min-w-12 bg-neutral-800 rounded-lg text-white font-medium px-3 focus:outline-neutral-400 focus:outline-2 inset-shadow-md " +
         (props.className || "")
       }
       value={props.value}