+++ changelog.md
@@ -1,3 +1,18 @@
+## 0.1.6 (NOT RELEASED)
+
+### [FEATURE]
+
+- Added settings
+- Added settings page
+- Added audio for learn and review
+
+### [REFACTOR]
+
+- Refactored some styles and now uses a global component for hanzi characters
+- Chagned the color of website
+
+## 0.1.5 (04/05/2025)
+
 ## 0.1.4 (04/05/2025)
 
 ### [FEATURE]